requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
click==8.1.7
dnspython==2.4.2
fastapi==0.105.0
h11==0.14.0
idna==3.6
motor==3.3.2
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.25.0
websockets==12.0
